Fix: multiple classes error
