+++ form_listener.js
@@ -1,0 +1,24 @@
+
+export default class FormListener {
+    constructor(bookManager){
+  
+      
+      this.bookManager = bookManager;
+      this.addEventListener();
+    }
+  
+    addEventListener=()=>{
+      const form = document.querySelector('.form');
+      form.addEventListener('submit', (event) => {
+          const formData = new FormData(form);
+          event.preventDefault();
+          const object = {};
+          formData.forEach((value, key) => {
+              object[key] = value;
+          });
+      
+          this.bookManager.add(object)
+      });
+    }
+  }
+
