+++ index.html
@@ -23,6 +23,6 @@
         <button  type="submit">Add</button>
 
     </form>
-    <script src="book_manager.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
